Updates from 4.2

+++ OpenGL/Texture.cpp
@@ -177,7 +177,7 @@ void Texture::LoadTextureArray(crossplatform::RenderPlatform* r, const std::vect
 	
 	width		= loadedTextures[0].x;
 	length		= loadedTextures[0].y;
-	arraySize	= loadedTextures.size();
+	arraySize	= (int)loadedTextures.size();
 	mips		= std::min(m,1 + int(floor(log2(width >= length ? width : length))));
 	dim			= 2;
 	depth		= 1;
